Fix: Correct data type for take_profit_price in Trade model
This commit resolves a persistent data integrity issue by correcting the SQLAlchemy model definition for the `trades` table.

The `take_profit_price` field was incorrectly defined as `Float`, while all other price-related fields were correctly defined as `Numeric`. This inconsistency was the root cause of unpredictable type conversion errors, leading to calculation failures for planned metrics.

By changing `take_profit_price` to `Numeric`, this commit ensures data type consistency across all price fields, eliminating the underlying cause of the bug and improving the overall stability and correctness of the data model.

+++ backend/app/Models/trade.py
@@ -105,7 +105,7 @@ class Trade(Base):
 
     # Optional fields
     stop_loss_price: Mapped[Optional[Numeric]] = mapped_column(Numeric, nullable=True)
-    take_profit_price: Mapped[Optional[float]] = mapped_column(Float, nullable=True)
+    take_profit_price: Mapped[Optional[Numeric]] = mapped_column(Numeric, nullable=True)
     lowest_price_during_trade: Mapped[Optional[Numeric]] = mapped_column(
         Numeric, nullable=True
     )
